Add files via upload
The repository contains .R and . Python files which contain the project code. The project was assigned and done under the supervision of Edwisor. The dataset provided is train_cab.csv and test_csv. 
CabFarePrediction_Report is a compiled pdf file of the project which consists of the different data science and machine learning algorithms that were used in the project and the reason for the same.